[IMP] changes in attendance report.
