+++ wizard/ng_church_collections_wizard_view.xml
@@ -36,7 +36,7 @@
                             <field name="date_to" required="1"/>
                         </group>
                         <group>
-                            <field name="attendance" class="oe_center" width="100%"/>
+                            <field name="program_id" class="oe_center" width="100%" options='{"no_open": True, "no_create": True}'/>
                         </group>
                     </group>
                     <footer>
